修复奇怪的bug

+++ include/PowDev.h
@@ -49,7 +49,7 @@ typedef union _MotorStatus_Typedef {
 class PowDevClass {
 public:
 	static uint8_t Status[3];
-	bool Valve(uint32_t status);
+	bool Valve(uint8_t status);
 	bool Valve(ValveCh_Typedef ch, FunctionalState NewState);
 	bool Motor(uint8_t status);
 	bool Motor(MotorCh_Typedef ch, uint8_t ctl);
